SH-344: fix originate calls not tracking callers
The on_state_change makes the decision to call either on_a_dial or
on_b_dial based on state of the channel before and after the state
change. Instead of DOWN to UP triggering on_a_dial it should trigger
on_b_dial. The confusion about that situation also showed in the
docstring, which is now rectified.